+++ desarrollo/src/main/java/com/example/desarrollo/model/Habitacion.java
@@ -20,8 +20,8 @@ public class Habitacion {
     private Long id;
 
     private String direccion;
-    private Long longitud;
-    private Long latitud;
+    private Double longitud;
+    private Double latitud;
     private double precio;
     private Integer area;
